Refactor unit tests for groups, users, and evaluation functionalities; consolidate tests into unified files, remove deprecated tests, and enhance test coverage for user and group management features.
