+++ apcpro-api/.delete-list.txt
@@ -1,0 +1,4 @@
+tests/unit/grupos.test.ts
+test-delete-api.js
+test-repository-methods.js
+teste-grupos.js
