--- apcpro-api/tests/unit/avaliacao-service-completo.test.ts
@@ -1,406 +1,0 @@
-/**
- * Testes Unitários - Avaliação Service Completo
- * Testa métodos     test('deve calcular RCQ quando for    test('deve lidar com dados mínimos obrigatórios', () => {
-      // Arrange - Dados mínimos
-      const input = {
-        peso: 70,
-        altura: 175,
-        idade: 25,
-        sexo: 'masculino',
-        tronco: {
-          cintura: 85,
-          quadril: 95
-        },
-        parteSuperior: {}
-      };
-
-      // Actcunferências', () => {
-      // Arrange - Com dados completos para RCQ
-      const input = {
-        peso: 70,
-        altura: 175,
-        idade: 25,
-        sexo: 'masculino',
-        tronco: {
-          cintura: 85,
-          quadril: 95
-        },
-        parteSuperior: {
-          pescoco: 38
-        }
-      };
-
-      // Acte avaliação e estrutura JSON para banco
- */
-
-import { AvaliacaoService } from '../../src/services/avaliacao-service';
-import { MedidasInput } from '../../src/utils/avaliacaoMedidas';
-import { Genero } from '../../src/models/genero-model';
-
-describe('AvaliacaoService - Estrutura JSON', () => {
-  let avaliacaoService: AvaliacaoService;
-
-  beforeEach(() => {
-    avaliacaoService = new AvaliacaoService();
-  });
-
-  describe('calcularIndices', () => {
-    test('deve calcular índices com dados válidos', () => {
-      // Arrange - Usando estrutura JSON esperada pelo processarMedidas
-      const input = {
-        peso: 70,
-        altura: 175,
-        idade: 25,
-        sexo: 'masculino',
-        tronco: {
-          cintura: 85,
-          quadril: 95
-        },
-        parteSuperior: {
-          pescoco: 38
-        }
-      };
-
-      // Act
-      const resultado = avaliacaoService.calcularIndices(input);
-
-      // Assert
-      expect(resultado).toBeDefined();
-      expect(typeof resultado).toBe('object');
-      expect(resultado).toHaveProperty('imc');
-      expect(resultado).toHaveProperty('classificacaoIMC');
-      expect(resultado.imc).toBeCloseTo(22.86, 2);
-      expect(resultado.classificacaoIMC).toBe('Normal');
-    });
-
-    test('deve calcular índices para mulher', () => {
-      // Arrange - Usando estrutura JSON esperada
-      const input = {
-        peso: 60,
-        altura: 165,
-        idade: 30,
-        sexo: 'feminino',
-        tronco: {
-          cintura: 68,
-          quadril: 95
-        },
-        parteSuperior: {
-          pescoco: 32
-        }
-      };
-
-      // Act
-      const resultado = avaliacaoService.calcularIndices(input);
-
-      // Assert
-      expect(resultado).toBeDefined();
-      expect(resultado).toHaveProperty('imc');
-      expect(resultado).toHaveProperty('classificacaoIMC');
-      expect(resultado.imc).toBeCloseTo(22.04, 2);
-      expect(resultado.classificacaoIMC).toBe('Normal');
-    });
-
-    test('deve calcular RCQ quando fornecidas circunferências', () => {
-      // Arrange - Com dados completos para RCQ
-      const input = {
-        peso: 70,
-        altura: 175,
-        idade: 25,
-        sexo: 'masculino',
-        tronco: {
-          cintura: 85,
-          quadril: 98
-        },
-        parteSuperior: {
-          pescoco: 38
-        }
-      };
-
-      // Act
-      const resultado = avaliacaoService.calcularIndices(input);
-
-      // Assert
-      expect(resultado).toBeDefined();
-      expect(resultado).toHaveProperty('rcq');
-      expect(resultado).toHaveProperty('classificacaoRCQ');
-      expect(resultado.rcq).toBeCloseTo(0.87, 2);
-      expect(resultado.classificacaoRCQ).toBe('Baixo risco'); // Corrigido para valor real retornado
-    });
-
-    test('deve lidar com dados mínimos obrigatórios', () => {
-      // Arrange - Dados mínimos necessários
-      const input = {
-        peso: 65,
-        altura: 170,
-        idade: 25,
-        sexo: 'masculino',
-        tronco: {
-          cintura: 80,
-          quadril: 90
-        },
-        parteSuperior: {}
-      };
-
-      // Act
-      const resultado = avaliacaoService.calcularIndices(input);
-
-      // Assert
-      expect(resultado).toBeDefined();
-      expect(resultado).toHaveProperty('imc');
-      expect(resultado).toHaveProperty('classificacaoIMC');
-      expect(resultado.imc).toBeCloseTo(22.49, 2);
-      expect(resultado.classificacaoIMC).toBe('Normal');
-    });
-
-    test('deve tratar dados inválidos', () => {
-      // Arrange - Dados inválidos mas no formato JSON esperado
-      const input = {
-        peso: -10,
-        altura: 0,
-        idade: 25,
-        sexo: 'masculino',
-        tronco: {
-          cintura: 80,
-          quadril: 90
-        },
-        parteSuperior: {}
-      };
-
-      // Act & Assert - Como o service pode tratar valores inválidos internamente,
-      // vamos verificar se retorna resultado ou lança erro
-      try {
-        const resultado = avaliacaoService.calcularIndices(input);
-        // Se não lança erro, pelo menos deve retornar um objeto válido
-        expect(resultado).toBeDefined();
-        expect(typeof resultado).toBe('object');
-      } catch (error) {
-        // Se lança erro, deve ser um erro relacionado a dados inválidos
-        expect(error).toBeInstanceOf(Error);
-      }
-    });
-  });
-
-  describe('calcularIndicesDirecto - Estrutura para JSON', () => {
-    test('deve calcular índices direto com dados válidos', () => {
-      // Arrange
-      const input: MedidasInput = {
-        peso: 80,
-        altura: 180,
-        genero: Genero.Masculino,
-        idade: 28,
-        cintura: 85
-      };
-
-      // Act
-      const resultado = avaliacaoService.calcularIndicesDirecto(input);
-
-      // Assert
-      expect(resultado).toBeDefined();
-      expect(resultado).toHaveProperty('imc');
-      expect(resultado).toHaveProperty('classificacaoIMC');
-      expect(resultado.imc).toBeCloseTo(24.69, 2);
-      expect(resultado.classificacaoIMC).toBe('Normal');
-    });
-
-    test('deve calcular com dobras cutâneas', () => {
-      // Arrange
-      const input: MedidasInput = {
-        peso: 75,
-        altura: 180,
-        genero: Genero.Masculino,
-        idade: 28,
-        cintura: 85,
-        dobras: {
-          triceps: 10,
-          subescapular: 12,
-          suprailiaca: 15
-        }
-      };
-
-      // Act
-      const resultado = avaliacaoService.calcularIndicesDirecto(input);
-
-      // Assert
-      expect(resultado).toBeDefined();
-      expect(resultado).toHaveProperty('imc');
-      expect(resultado).toHaveProperty('percentualGC_Pollock');
-      expect(resultado.imc).toBeCloseTo(23.15, 2);
-      expect(resultado.percentualGC_Pollock).toBeDefined();
-    });
-
-    test('deve retornar null para dobras quando não fornecidas', () => {
-      // Arrange
-      const input: MedidasInput = {
-        peso: 70,
-        altura: 175,
-        genero: Genero.Masculino,
-        idade: 25,
-        cintura: 85
-      };
-
-      // Act
-      const resultado = avaliacaoService.calcularIndicesDirecto(input);
-
-      // Assert
-      expect(resultado).toBeDefined();
-      expect(resultado.percentualGC_Pollock).toBeNull();
-      expect(resultado.classificacaoGC_Pollock).toBeNull();
-    });
-
-    test('deve retornar objeto que pode ser serializado como JSON', () => {
-      // Arrange
-      const input: MedidasInput = {
-        peso: 70,
-        altura: 175,
-        genero: Genero.Masculino,
-        idade: 25,
-        cintura: 85,
-        quadril: 98
-      };
-
-      // Act
-      const resultado = avaliacaoService.calcularIndicesDirecto(input);
-
-      // Assert
-      expect(resultado).toBeDefined();
-      
-      // Verifica se pode ser serializado como JSON
-      const jsonString = JSON.stringify(resultado);
-      expect(jsonString).toBeDefined();
-      expect(typeof jsonString).toBe('string');
-      
-      // Verifica se pode ser deserializado
-      const objetoDeserializado = JSON.parse(jsonString);
-      expect(objetoDeserializado).toEqual(resultado);
-    });
-
-    test('deve manter estrutura consistente para salvar no banco', () => {
-      // Arrange
-      const input: MedidasInput = {
-        peso: 75,
-        altura: 180,
-        genero: Genero.Masculino,
-        idade: 28,
-        cintura: 85,
-        quadril: 98,
-        dobras: {
-          triceps: 10,
-          subescapular: 12,
-          suprailiaca: 15
-        }
-      };
-
-      // Act
-      const resultado = avaliacaoService.calcularIndicesDirecto(input);
-
-      // Assert - Verifica se tem os campos esperados para salvar como JSON
-      expect(resultado).toBeDefined();
-      expect(resultado).toHaveProperty('imc');
-      expect(resultado).toHaveProperty('classificacaoIMC');
-      expect(resultado).toHaveProperty('rcq');
-      expect(resultado).toHaveProperty('classificacaoRCQ');
-      expect(resultado).toHaveProperty('percentualGC_Pollock');
-      expect(resultado).toHaveProperty('classificacaoGC_Pollock');
-      
-      // Verifica se é um objeto válido para JSON
-      expect(typeof resultado).toBe('object');
-      expect(resultado).not.toBeNull();
-    });
-
-    test('deve ser compatível com estrutura de avaliação no banco', () => {
-      // Arrange - Simula dados que seriam enviados para salvar
-      const input: MedidasInput = {
-        peso: 70,
-        altura: 175,
-        genero: Genero.Masculino,
-        idade: 25,
-        cintura: 85,
-        quadril: 98,
-        pescoco: 38
-      };
-
-      // Act
-      const resultado = avaliacaoService.calcularIndicesDirecto(input);
-
-      // Assert - Verifica se o resultado pode ser usado como campo 'resultado' da avaliação
-      const avaliacaoParaSalvar = {
-        tipo: 'medidas',
-        status: 'pendente',
-        resultado: resultado, // Este objeto deve ser válido para JSON
-        validadeAte: null,
-        objetivoClassificado: null
-      };
-
-      // Testa serialização JSON
-      expect(() => JSON.stringify(avaliacaoParaSalvar)).not.toThrow();
-      
-      const jsonSerializado = JSON.stringify(avaliacaoParaSalvar);
-      const objetoDeserializado = JSON.parse(jsonSerializado);
-      
-      expect(objetoDeserializado.resultado).toBeDefined();
-      expect(objetoDeserializado.resultado.imc).toBeDefined();
-      expect(objetoDeserializado.resultado.classificacaoIMC).toBeDefined();
-    });
-
-    test('deve preservar estrutura original dos dados de entrada', () => {
-      // Arrange
-      const input: MedidasInput = {
-        peso: 75,
-        altura: 180,
-        genero: Genero.Masculino,
-        idade: 28,
-        cintura: 85,
-        quadril: 98
-      };
-
-      // Act
-      const resultado = avaliacaoService.calcularIndicesDirecto(input);
-
-      // Assert - Verifica se mantém dados originais + cálculos
-      expect(resultado).toHaveProperty('peso', input.peso);
-      expect(resultado).toHaveProperty('altura', input.altura);
-      expect(resultado).toHaveProperty('genero', input.genero);
-      expect(resultado).toHaveProperty('idade', input.idade);
-      
-      // E também os índices calculados
-      expect(resultado).toHaveProperty('imc');
-      expect(resultado).toHaveProperty('classificacaoIMC');
-      expect(resultado).toHaveProperty('rcq');
-      expect(resultado).toHaveProperty('classificacaoRCQ');
-    });
-
-    test('deve validar que resultado JSON contém informações necessárias', () => {
-      // Arrange
-      const input: MedidasInput = {
-        peso: 70,
-        altura: 175,
-        genero: Genero.Masculino,
-        idade: 25,
-        cintura: 85
-      };
-
-      // Act
-      const resultado = avaliacaoService.calcularIndicesDirecto(input);
-
-      // Assert - Verifica estrutura específica para JSON no banco
-      expect(resultado).toBeDefined();
-      
-      // Campos básicos obrigatórios
-      expect(resultado).toHaveProperty('imc');
-      expect(typeof resultado.imc).toBe('number');
-      
-      expect(resultado).toHaveProperty('classificacaoIMC');
-      expect(typeof resultado.classificacaoIMC).toBe('string');
-      
-      // Verifica que não há propriedades undefined
-      const resultadoKeys = Object.keys(resultado);
-      resultadoKeys.forEach(key => {
-        expect(resultado[key as keyof typeof resultado]).not.toBeUndefined();
-      });
-      
-      // Verifica que JSON parse/stringify mantém integridade
-      const jsonRoundTrip = JSON.parse(JSON.stringify(resultado));
-      expect(jsonRoundTrip).toEqual(resultado);
-    });
-  });
-});
