--- apcpro-api/tests/unit/avaliacao-service.test.ts
@@ -1,33 +1,0 @@
-// tests/unit/avaliacao-service.test.ts
-import { AvaliacaoService } from '../../src/services/avaliacao-service';
-
-describe('AvaliacaoService', () => {
-  describe('Instanciação básica', () => {
-    it('deve instanciar sem erros', () => {
-      expect(() => new AvaliacaoService()).not.toThrow();
-    });
-
-    it('deve ter estrutura básica', () => {
-      const service = new AvaliacaoService();
-      expect(service).toBeDefined();
-      expect(service).toBeInstanceOf(AvaliacaoService);
-    });
-  });
-
-  describe('Validações básicas', () => {
-    let service: AvaliacaoService;
-
-    beforeEach(() => {
-      service = new AvaliacaoService();
-    });
-
-    it('deve ser definido após instanciação', () => {
-      expect(service).toBeDefined();
-    });
-
-    it('deve ter propriedades básicas', () => {
-      expect(service).toBeDefined();
-      expect(service.constructor.name).toBe('AvaliacaoService');
-    });
-  });
-});
