--- apcpro-api/tests/unit/avaliacoes-controller.test.ts
@@ -1,87 +1,0 @@
-/**
- * Testes Unitários - Avaliação Controller (Simplificado)
- * Testa as rotas e validações do controller de avaliações
- */
-
-import { Request, Response, NextFunction } from 'express';
-
-describe('AvaliacaoController - Testes Simples', () => {
-  let mockRequest: Partial<Request>;
-  let mockResponse: Partial<Response>;
-  let mockNext: NextFunction;
-
-  beforeEach(() => {
-    // Mock do request e response
-    mockRequest = {
-      body: {},
-      params: {},
-      query: {}
-    };
-    
-    mockResponse = {
-      status: jest.fn().mockReturnThis(),
-      json: jest.fn().mockReturnThis(),
-      send: jest.fn().mockReturnThis()
-    };
-
-    mockNext = jest.fn();
-  });
-
-  describe('Estrutura e Tipagem', () => {
-    test('deve ter Request, Response e NextFunction tipados corretamente', () => {
-      expect(mockRequest).toBeDefined();
-      expect(mockResponse).toBeDefined();
-      expect(mockNext).toBeDefined();
-      
-      // Verificar se as funções de response estão mockadas
-      expect(typeof mockResponse.status).toBe('function');
-      expect(typeof mockResponse.json).toBe('function');
-      expect(typeof mockNext).toBe('function');
-    });
-
-    test('deve processar body da requisição', () => {
-      mockRequest.body = {
-        peso: 70,
-        altura: 175
-      };
-
-      expect(mockRequest.body.peso).toBe(70);
-      expect(mockRequest.body.altura).toBe(175);
-    });
-  });
-
-  describe('Validação de Dados', () => {
-    test('deve aceitar dados válidos de medidas básicas', () => {
-      const dadosValidos = {
-        peso: 70,
-        altura: 175,
-        genero: 'masculino',
-        idade: 25
-      };
-
-      mockRequest.body = dadosValidos;
-
-      expect(mockRequest.body.peso).toBeGreaterThan(0);
-      expect(mockRequest.body.altura).toBeGreaterThan(0);
-      expect(['masculino', 'feminino']).toContain(mockRequest.body.genero);
-      expect(mockRequest.body.idade).toBeGreaterThan(0);
-    });
-
-    test('deve identificar dados inválidos', () => {
-      const dadosInvalidos = {
-        peso: -10,
-        altura: 0,
-        genero: 'invalido',
-        idade: -5
-      };
-
-      mockRequest.body = dadosInvalidos;
-
-      // Validações básicas
-      expect(mockRequest.body.peso).toBeLessThanOrEqual(0);
-      expect(mockRequest.body.altura).toBeLessThanOrEqual(0);
-      expect(['masculino', 'feminino']).not.toContain(mockRequest.body.genero);
-      expect(mockRequest.body.idade).toBeLessThanOrEqual(0);
-    });
-  });
-});
